requirements.txt
# Основные зависимости для анализа данных
pandas>=1.5.0
numpy>=1.21.0
jupyter>=1.0.0
notebook>=6.4.0

# Веб-скрапинг и парсинг
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Обработка естественного языка
spacy>=3.5.0

# Работа с OpenAI API
openai>=1.0.0
httpx>=0.25.0

# Работа с переменными окружения
python-dotenv>=0.19.0

# Работа с конфигурацией
PyYAML>=6.0

# GUI (если понадобится)
PyQt5>=5.15.0

# Работа с Excel файлами
openpyxl>=3.0.0
xlrd>=2.0.0

# Дополнительные утилиты
tqdm>=4.64.0

# Тестирование
pytest>=7.0.0
pytest-cov>=4.0.0

# Дополнительные зависимости для веб-скрапинга
urllib3>=1.26.0
certifi>=2021.5.30

# Генерация .apkg для Anki
genanki>=0.13.0

# Интеграция с Anki
anki>=2.1.0
py-ankiconnect>=1.1.0

# Аудио (необязательно): фоновая 8-бит музыка для CLI
sounddevice>=0.4.6
